fix: 7주차 피드백 반영
