+++ Week7/app/src/main/java/com/example/week7/navigation/NavigationItem.kt
@@ -1,0 +1,7 @@
+package com.example.week7.navigation
+
+data class NavigationItem(
+    val title: String,
+    val route: Any,
+    val iconRes: Int
+)
